Remove EDH feeds api methods & tests
